backend/lambda/query-translation-status/requirements.txt
psycopg2-binary==2.9.10
boto3==1.35.90
